+++ src/resolvers/Mutation/post.ts
@@ -1,6 +1,8 @@
 import { sendPostNotificationsAsync } from "../../communications/notifications";
 import { MutationResolvers } from "../../generated/graphqlgen";
-import { AuthError, getPersonId } from "../../utils";
+
+import { AuthError } from '../../errors'
+import { getPersonId } from "../../utils";
 
 export const post: Pick<MutationResolvers.Type, "createPost" | "editPost" | "deletePost"> = {
   createPost: async (parent, { threadId, content }, ctx, info) => {
