+++ src/resolvers/Mutation/thread.ts
@@ -1,6 +1,7 @@
 import { sendPostNotificationsAsync } from "../../communications/notifications";
 import { MutationResolvers } from "../../generated/graphqlgen";
-import { checkGroupMembership, getPersonId } from "../../utils";
+
+import { isMember, getPersonId } from "../../utils";
 
 export const thread: Pick<
   MutationResolvers.Type,
@@ -8,7 +9,7 @@ export const thread: Pick<
 > = {
   createThread: async (parent, { groupId, title, content }, ctx, info) => {
     const personId = getPersonId(ctx);
-    await checkGroupMembership(ctx, groupId);
+    await isMember(ctx, groupId);
 
     const newThread = await ctx.prisma.createThread({
       title,
@@ -50,7 +51,7 @@ export const thread: Pick<
       })
       .group()
       .id();
-    await checkGroupMembership(ctx, groupId);
+    await isMember(ctx, groupId);
     return await ctx.prisma.updateThread({
       where: {
         id: threadId
@@ -68,7 +69,7 @@ export const thread: Pick<
       })
       .group()
       .id();
-    await checkGroupMembership(ctx, groupId);
+    await isMember(ctx, groupId);
     await ctx.prisma.deleteThread({
       id: threadId
     });
@@ -86,7 +87,7 @@ export const thread: Pick<
       })
       .group()
       .id();
-    checkGroupMembership(ctx, groupId);
+    isMember(ctx, groupId);
     const prevData = await ctx.prisma.thread({ id: threadId });
     if (!prevData) {
       throw new Error("Thread does not exist.");
